refactor: HeroCard
